refactor(nav): extract useLocaleSwitcher and clean up preferences block
Post-/simplify pass:
- Extract useLocaleSwitcher hook to share locale-switching logic
  between LocaleToggle and the new LocaleSegmented in UserMenu
- Fix i18n leak: segmented controls aria-label now read from
  UserMenu.preferences (was hardcoded "Locale"/"Theme")
- Replace the plain div wrapper around the mobile preferences block
  with DropdownMenuGroup, keeping Radix focus model consistent
- Drop lightLabel/darkLabel prop drilling on ThemeSegmented (it now
  reads its own translations, mirroring LocaleSegmented)
